Initial commit for ruby gem.
