+++ ruby/Gemfile
@@ -1,0 +1,4 @@
+source "http://rubygems.org"
+
+gemspec
+
