+++ ruby/Rakefile
@@ -1,0 +1,7 @@
+require 'bundler/gem_tasks'
+require 'rspec/core/rake_task'
+
+desc 'Run Specs'
+RSpec::Core::RakeTask.new(:spec)
+
+task :default => :spec
