+++ ruby/lib/bitcoin_addrgen/version.rb
@@ -1,0 +1,3 @@
+module BitcoinAddrgen
+  VERSION = "0.0.1"
+end
