+++ ruby/spec/spec_helper.rb
@@ -1,0 +1,6 @@
+require 'rubygems'
+require 'rspec'
+
+$LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
+$LOAD_PATH.unshift(File.dirname(__FILE__))
+require 'bitcoin-addrgen'
